+++ pkg/schemas/generators/hmac.go
