+++ pkg/schemas/generators/secret_entry.go
