+++ pkg/schemas/generators/types.go
