+++ pkg/schemas/helpers.go
@@ -1,4 +1,4 @@
-package schema
+package schemas
 
 import (
 	"fmt"
